requirements.txt
flask
flask-admin
flask-sqlalchemy
jinja2
fs
Pillow
